BACKEND/requirements.txt
Django==2.2
djangorestframework==3.9.1

